+++ app/routers/financial_categories.py
